Modularized the Vertex logic
